# src/server/routes/swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "version": "1.0.0",
        "title": "UTD Degree Planner and Auditor",
        "description": "API Documentation for UTD Degree Planner and Auditor"
    },
    "servers": [
        {
            "url": "http://localhost:3000/"
        }
    ],
...
    "paths": {
...
        "/api/testAndTransferCredits/transferCreditEquivalenciesByUTDCourse": {
            "get": {
                "description": "Get transfer credit equivalencies by UTD course",
                "summary": "Get transfer credit equivalencies by UTD course",
                "parameters": [
                    {
                        "schema": {
                            "type": "string",
                            "minLength": 1
                        },
                        "required": true,
                        "name": "prefix",
                        "in": "query"
                    },
                    {
                        "schema": {
                            "type": "string",
                            "minLength": 4,
                            "maxLength": 4
                        },
                        "required": true,
                        "name": "number",
                        "in": "query"
                    },
                    {
                        "schema": {
                            "type": "string"
                        },
                        "required": false,
                        "name": "schoolsFilter",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/testAndTransferCredits/transferCreditEquivalenciesByTransferSchool": {
            "get": {
                "description": "Get transfer credit equivalencies by transfer school",
                "summary": "Get transfer credit equivalencies by transfer school",
                "parameters": [
                    {
                        "schema": {
                            "type": "string",
                            "minLength": 1
                        },
                        "required": true,
                        "name": "transferSchoolSchoolID",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}